Bumped packages to fix image handling in ios18

+++ SmallPhotos.Service.Tests/FluentAssertionExtensions.cs
@@ -11,8 +11,8 @@ public static class FluentAssertionExtensions
     public static AndConstraint<GenericCollectionAssertions<byte>> BeOfSize(this GenericCollectionAssertions<byte> imageBytes, Size expectedImageSize)
     {
         using MagickImage image = new(imageBytes.Subject.ToArray());
-        image.Width.Should().Be(expectedImageSize.Width);
-        image.Height.Should().Be(expectedImageSize.Height);
+        image.Width.Should().Be((uint)expectedImageSize.Width);
+        image.Height.Should().Be((uint)expectedImageSize.Height);
         return new(imageBytes);
     }
 }
